refact: reestructure pages
